No Chains; Segments stored by offset (#28)
Workstation deprecates "Chain" under the hood entirely, persists Segments by offset (integer) only
https://www.pivotaltracker.com/story/show/186004996

+++ nexus/src/main/java/io/xj/nexus/work/CraftWork.java
@@ -99,7 +99,7 @@ public interface CraftWork {
    @param offset of segment
    @return the segment at the given offset
    */
-  Optional<Segment> getSegmentAtOffset(long offset);
+  Optional<Segment> getSegmentAtOffset(int offset);
 
   /**
    Get the segments spanning the given time range, if they are ready- if not, return an empty list
